requirements.txt
tensorflow==2.16.1
keras==3.3.3
numpy==1.26.4
pillow==10.3.0
matplotlib==3.8.4
scikit-learn==1.4.2
tk